normal monster all of implementation

+++ WOL/WOL/timer.cpp
@@ -99,5 +99,5 @@ float Timer::delta()
    // return (1000.f / fps) * 0.001 * time_scale;
 	//return 0.016f;
 
-	return dt /1000.f * time_scale;
+	return (float)dt /1000.f * (float)time_scale;
 }
